Move ;xkcd to fun cog
