+++ roxbot/load_config.py
@@ -37,6 +37,5 @@ cogs = [
 	"roxbot.cogs.twitch",
 	"roxbot.cogs.util",
 	"roxbot.cogs.voice",
-    "roxbot.cogs.xkcd"
 	#"roxbot.cogs.gss"
 ]
